re-structure utils as modules for tree shaking
