+++ src/types.ts
@@ -1,0 +1,35 @@
+import { BaseProvider } from '@ethersproject/providers';
+
+export interface Spec {
+  getMetadata: (
+    provider: BaseProvider,
+    ownerAddress: string | undefined | null,
+    contractAddress: string,
+    tokenID: string,
+    options?: AvatarResolverOpts
+  ) => Promise<any>;
+}
+
+export interface AvatarResolverOpts {
+  cache?: number;
+  ipfs?: string;
+  arweave?: string;
+  urlDenyList?: string[];
+}
+
+export interface AvatarRequestOpts {
+  jsdomWindow?: any;
+}
+
+export type Gateways = {
+  ipfs?: string;
+  arweave?: string;
+};
+
+export interface ImageURIOpts {
+  metadata: any;
+  customGateway?: string;
+  gateways?: Gateways;
+  jsdomWindow?: any;
+  urlDenyList?: string[];
+}
