+++ src/utils/assert.ts
@@ -1,0 +1,6 @@
+// simple assert without nested check
+export function assert(condition: any, message: string) {
+  if (!condition) {
+    throw message;
+  }
+}
